ssa-example/requirements.txt
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dotenv==0.9.9
idna==3.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
